Moved backend/hardware/__init__.py to backend/hardware/core.py
