+++ backend/tests/tasks_test.py
@@ -2,7 +2,7 @@ import recipes
 from recipes import tasks
 import pytest
 import yaml
-import hardware
+import hardware.core
 from unittest.mock import patch, MagicMock
 
 
@@ -145,7 +145,7 @@ def devices(request):
 
 @pytest.fixture
 def microlab(request, devices):
-    return hardware.MicroLabHardware(devices)
+    return hardware.core.MicroLabHardware(devices)
 
 
 # HEATING
